fix: reduce extruded amount by departure width, not arrival width
the extrusion process starts by putting down a circle during the unretraction and then puts down moon shapes during extrusion moves.
This means that the extrusion lines before a short wide segment are wider instead of the extrusion lines after.
Because there is a small delay in extrusion, this might be a better choice